[recovery]: Fixed the comparision of error codes from runner.
When we compare the error code returned by the runner with integer
literals, we need to convert the integer literal to int32 type.

This commit makes this change for the error code in exec for
re-flashing the CR50 firmware.

BUG=b:254324617
TEST=build, go test ./...

Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3964319
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#52310}

+++ go/src/infra/cros/recovery/internal/execs/cros/cros_cr50_execs.go
@@ -129,7 +129,11 @@ func reflashCr50FwExec(ctx context.Context, info *execs.ExecInfo) error {
 		if !ok {
 			return errors.Annotate(err, "reflash cr50 fw: cannot find error code").Err()
 		}
-		if errorCode != 1 {
+		// The errorCode value stored in the empty interface in the
+		// error tag is of type int32. To compare it with an integer
+		// literal, we need to convert the literal into int32 value,
+		// otherwise the comparison will always return 'false'.
+		if errorCode != int32(1) {
 			return errors.Annotate(err, "reflash cr50 fw: fail to flash %q", info.GetChromeos().GetCr50Phase()).Err()
 		}
 	}
